Refactor iframe integration page and add Matomo tracking
- Merged `/iframe.vue` content into `/integrer-nos-simulateurs.vue`
- Removed redundant `/iframe.vue` page
- Added Matomo and `install` dependencies to `package.json`
- Minor code cleanup in `Survey.vue`
