+++ plugins/matomo.client.js
@@ -1,0 +1,16 @@
+import { defineNuxtPlugin } from '#app'
+import VueMatomo from 'vue-matomo'
+
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.vueApp.use(VueMatomo, {
+    host: 'https://stats.beta.gouv.fr/',
+    siteId: 199,
+    // Enables automatically registering pageviews on the router
+    router: nuxtApp.$router,
+    enableLinkTracking: true,
+    requireConsent: false,
+    trackInitialView: true,
+    disableCookies: true,
+    requireCookieConsent: false,
+  })
+})
